rskrza1: Disable PERF_EVENTS in order to allow GDB debugging
When CONFIG_PERF_EVENTS is enabled, the file hw_breakpoint.c is included
in the build. The issue is that when that driver is initialized, it clears
out all the HW breakpoints. That makes if very difficult to set breakpoints
in probe functions that you want to debug on boot.