rskrza1: add touchscreen support
The touchscreen on the RSK is connected to I2C ch0.
The interrupt line is connected to IRQ1.
Note that the ARM GIC interrupt controller driver cannot set the trigger
level of our hardware IRQ pins because that is controlled in a separate
HW block. Therefore we need to manually set that up in the board file.